+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.3.0 (2024-10-12)
+
+- Softened border colors
+
+- Changed accent color to purple
+
+- Desaturated background colors a little bit
+
+- Switched status bar to slime color
+
+- Changed tooltips to pale green
+
 ## 1.2.0 (2024-08-26)
 
 - Fixed the extension description text color